Add simple example .gitlab.yml file
This patch adds an example .gitlab.yml file to be used by the gitlab CI.

It should be very easily transferable to many projects that use docker
as their build system wrapper. It is not complete yet, it does not
handle the release stage, nor has it been adapted to other languages,
other then POSIX shell scripts.

It is however a very good template to get started from. Further more
these steps are best when also wrapped in a generic makefile/build
script as can be seen in jedi-system-update [0].

[0] https://gitlab.com/ultimaker/embedded/platform/jedi-system-update

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>